+++ packages/sdk-svelte/src/lib/index.ts
@@ -1,4 +1,3 @@
 /* eslint-disable unicorn/no-empty-file */
 
 export { default as RobotClient } from './robot-client/robot-client.svelte';
-export * from './robot-client/use-robot-client';
